backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
pillow==10.1.0
qrcode==7.4.2
redis==5.0.1
celery==5.3.4
email-validator==2.1.0
motor==3.3.2
pymongo==4.6.1 